Update imports after renames
