+++ codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/serviceclientconfiguration.java
@@ -10,7 +10,7 @@ import software.amazon.awssdk.core.client.config.SdkClientConfiguration;
 import software.amazon.awssdk.core.client.config.SdkClientOption;
 import software.amazon.awssdk.core.internal.SdkInternalAdvancedClientOption;
 import software.amazon.awssdk.endpoints.EndpointProvider;
-import software.amazon.awssdk.http.auth.spi.AuthSchemeProvider;
+import software.amazon.awssdk.http.auth.spi.scheme.AuthSchemeProvider;
 import software.amazon.awssdk.identity.spi.AwsCredentialsIdentity;
 import software.amazon.awssdk.identity.spi.IdentityProvider;
 import software.amazon.awssdk.regions.Region;
@@ -230,7 +230,7 @@ public final class JsonProtocolTestsServiceClientConfiguration extends AwsServic
                 return null;
             }
             return Validate.isInstanceOf(JsonProtocolTestsAuthSchemeProvider.class, result, "Expected an instance of "
-                    + JsonProtocolTestsAuthSchemeProvider.class.getSimpleName());
+                                                                                            + JsonProtocolTestsAuthSchemeProvider.class.getSimpleName());
         }
 
         @Override
